pyproject.toml
[project]
name = "crypto-proxy-service"
version = "0.1.0"
description = "HTTP and WebSocket proxy service skeleton"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Your Name" }]

dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "httpx>=0.27",
  "websockets>=13.1",
  "python-dotenv>=1.0",
]

[project.scripts]
crypto-proxy-service = "crypto_proxy_service.main:run"

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/crypto_proxy_service"]
